네이버

+++ src/main/java/com/sparta/WeatherWear/repository/WeatherRepository.java
@@ -4,5 +4,5 @@ package com.sparta.WeatherWear.repository;
 import com.sparta.WeatherWear.entity.Weather;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WeatherRepository extends JpaRepository<Weather, Integer> {
+public interface WeatherRepository extends JpaRepository<Weather, Long> {
 }
